requirement.txt
python-telegram-bot==20.3
gspread==6.1.4
oauth2client==4.1.3
pytz==2024.1